HoosHungryBackend/hooshungrybackend/pyproject.toml
[tool.poetry]
name = "hooshungrybackend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.2.6"
djangorestframework = "^3.16.1"
requests = "^2.32.5"
celery = "^5.5.3"
redis = "^7.0.1"
gunicorn = "^23.0.0"
django-cors-headers = "^4.9.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
